doc: getting started - ZBD emu null_blk 2 of x
This PR significantly rewrites and reworks some of
the material in the "Zoned Block Device Emulation
with null_blk" chapter of the Getting Started
Guide. I have added some subsections to this
chapter in the interest of signposting so that
the reader's cognitive load is reduced. I have also
restated in the text some of the information
contained in the parameters and arguments of the
examples included in the text, because in the
earlier revision they might have been
too implicit for the first-time reader.

2021 Aug 02: I have included Damien's corrections
             and notes.

Signed-off-by: Zac Dover <zac.dover@gmail.com>