doc: nullblk.md - improve wording
s/method for creating/way to create/

Every time I re-read this material, I have a
better idea of what it means to communicate,
and every time I re-read this material I am
able to revise the English so that it sounds
more natural.

+ 3 other commits that make various improvements
to the flow and wording of the English text.

Signed-off-by: Zac Dover <zac.dover@gmail.com>

doc: nullblk.md - rewrite emu text

This PR rewrites the following sections of
nullblk.md:

  1. Emulating SMR HDD
  2. Emulating NVMe ZNS SSD

This PR makes minor syntactic changes, which
are intended to lighten the cognitive load of
the reader.

Signed-off-by: Zac Dover <zac.dover@gmail.com>

doc: improve libzbc in prerequisite.md

This PR makes a few syntactic changes to the
"libzbc" subsection of the "Prerequisites"
chapter of the Getting Started Guide.

The lack of relative clauses in this section
was bothering me, and it was a stone in my
shoe.

Signed-off-by: Zac Dover <zac.dover@gmail.com>

doc: smr-disk.md - "Checking the Disk Info'

This PR improves the flow of the English in
"Checking the Disk Info" in the "Serial ATA
ZAC Disks and SATA Host Controllers" section
of the "Getting Started with SMR Disks" chapter
of the Getting Started Guide.

Signed-off-by: Zac Dover <zac.dover@gmail.com>

(squash) Damien's notes.

Removing "ring buffer" from the mentions of
dmesg output.

Signed-off-by: Zac Dover <zac.dover@gmail.com>

(squash) ingesting material in PR#22

Because I tried (and failed) to cleverly
nest commits in my locally clone, I created
a spaghetti monster. This commit un-spaghettis
part of that tangle by rescuing the changes made
to the following section:

* HBA Compatibility

Signed-off-by: Zac Dover <zac.dover@gmail.com>

(squash) s/failsthe/fails the/

Geez.

Signed-off-by: Zac Dover <zac.dover@gmail.com>