doc: rewrite smr-disk.md top matter
This PR rewrites the overview text at the
top of the "Getting Started with SMR Disks"
chapter of the Getting Started Guide.

There are some phrases in this document that cry out
to me for revision ("LBA space zoning information"),
but experience has taught me to await further experience
before I begin revising this.

When I have finished revising this guide, I will no doubt
return to this text in order to revise it yet again.

All writing is rewriting. Einmal ist keinmal.

Signed-off-by: Zac Dover <zac.dover@gmail.com>