doc: update Getting Started index text
This PR makes minor improvements to the English
language on the index.md page. Now that I have
a better sense of the technology of zoned
namespaces than I had a month ago, I can more
confidently assert that the changes in this PR
reduce the ambiguity of the text and make that
text communicate more clearly.

This PR started as a correction of the noun "setup"
to the intended verb "to set up", but time and
my persnickety nature caused that initial correction
to expand.:

Signed-off-by: Zac Dover <zac.dover@gmail.com>