doc: update second half of dm.md
This updates the second half of the "Device Mapper page.

Most of the updates in this file are typical
present-participle-to-relative-clause changes, but there
are a couple of sentences in this half that require significant
rewriting. I plan to use this PR to have a discussion that will
result in that rewriting.

Signed-off-by: Zac Dover <zac.dover@gmail.com>