doc: rewrite fs.md up to i/o management
This PR rewrites the fs.md file up to (but
excluding) the "I/O Management" section.

I have chosed to break off here, because there
are two sentences in this section that require
elucidation and the I/O Management section will
require significant clarification. So I thought
that we would take care of this before continuing.

Signed-off-by: Zac Dover <zac.dover@gmail.com>