doc/tools: refine qemu.md
This PR refines the English in tools/qemu.md.

It also makes some structural changes, and includes
some new compositions of information. For this reason,
it should be carefully examined to make certain that I
have not introduced technical errors or misrepresentations
into the text.

This represents the first serious alteration to a section
of the Zoned Storage documentation that I have made. If it
is a success, future PRs will be more like this than like
the ones that preceded it.

Signed-off-by: Zac Dover <zac.dover@gmail.com>