doc: improve getting-started/smr-disk.md 1st Sect
This PR makes minor but crucial syntactical changes to
the English in the front matter of the "Getting Started
with SMR Hard-Disks" chapter of the Getting Started Guide.

It might be that this little section means to say:
1. Drive Managed Interfaces are irrelevant to ZonedStorage.
2. Host Aware devices will work with ZonedStorage, but
   ZonedStorage treats them the way that it treats Host
   Managed devices (which is to say that it does not support
   random write capabilities, but instead respects the strong
   sequential write constraints imposed by the Zoned Block
   Interface).

If this list of two things does indeed preserve the intent of
this section, then I might return in the future to recast
this information more pithily. I set this comment down in
print here as a note to a future version of me, who will be
equipped with a more complete understanding of the interface
implementations and the proper manner of their explanation to
the ignorant.

(Here's the message from a squashed commit)
(squash) Damien's notes

I normalized the indentation by making it all carriage
returns and single spaces instead of the thing that vim
had done when I Visually highlighed the text and ran "gq"
on it.

I also added a colon to a section heading in order to be
consistent with the heading in the other listitem in a list.

Signed-off-by: Zac Dover <zac.dover@gmail.com>

(squash) fixing a long line

Signed-off-by: Zac Dover <zac.dover@gmail.com>

(squash) wrangling whitespace

Signed-off-by: Zac Dover <zac.dover@gmail.com>