doc/intro: improve prepositions in "Principle"
This second, later pass over the Introduction (in preparation
for the reorganization of the documentation) makes it possible
for me to see with greater confidence that some of the pre-
positions in this section were not what they should be.

Signed-off-by: Zac Dover <zac.dover@gmail.com>

doc/intro: intro.md - rewrite rest of "Principle"

This PR rewrites the part of the section "Principle" in
the introduction that I did not rewrite in the other pull
reqeust that I made today. After a bit of thinking about
the English grammar that was in the text a few minutes ago,
I decided that it could be made clearer.

Signed-off-by: Zac Dover <zac.dover@gmail.com>

doc/intro: add "sequential write constraint" def.

This adds an explicit identification and definition of the
"sequential write constraint" in the "Principle" section
of the Zoned Storage Devices chapter of the Introduction.

The "sequential write constraint" is referred to very fre-
quently in the ZonedStorage documentation, but there is, as
far as I can recall, no explicit mention of the constraint.

This PR remedies that.

Signed-off-by: Zac Dover <zac.dover@gmail.com>

doc/intro/zoned-storage: rewrite L. eco. support

This PR improves the English in the "Linux Ecosystem Support"
section of the "Zoned Storage" intro.

I have been carefully reading Lydia Davis's notes to her trans-
lation of Proust's "À la recherche du temps perdu", and it has
given me a better sense of prepositions are they are translated
from French to English.

My improved (as compared to last year) understanding of Zoned-
Storage makes me confident that the changes in this PR improve
the readability and accuracy of the English in this section.

Signed-off-by: Zac Dover <zac.dover@gmail.com>

doc/intro: refine "Devel. for ZS"

This PR refines the English in the "Developing for Zoned
Storage" section of the Introduction page.

A year's familiarity with ZonedStorage has increased my confidence
that the changes to the sentences in this PR improve the readability
and technical accuracy of this section.

Signed-off-by: Zac Dover <zac.dover@gmail.com>