doc/intro: intro.md - rewrite rest of "Principle"
This PR rewrites the part of the section "Principle" in
the introduction that I did not rewrite in the other pull
reqeust that I made today. After a bit of thinking about
the English grammar that was in the text a few minutes ago,
I decided that it could be made clearer.

Signed-off-by: Zac Dover <zac.dover@gmail.com>