doc/intro: add "sequential write constraint" def.
This adds an explicit identification and definition of the
"sequential write constraint" in the "Principle" section
of the Zoned Storage Devices chapter of the Introduction.

The "sequential write constraint" is referred to very fre-
quently in the ZonedStorage documentation, but there is, as
far as I can recall, no explicit mention of the constraint.

This PR remedies that.

Signed-off-by: Zac Dover <zac.dover@gmail.com>