doc/intro/zoned-storage: rewrite L. eco. support
This PR improves the English in the "Linux Ecosystem Support"
section of the "Zoned Storage" intro.

I have been carefully reading Lydia Davis's notes to her trans-
lation of Proust's "À la recherche du temps perdu", and it has
given me a better sense of prepositions are they are translated
from French to English.

My improved (as compared to last year) understanding of Zoned-
Storage makes me confident that the changes in this PR improve
the readability and accuracy of the English in this section.

Signed-off-by: Zac Dover <zac.dover@gmail.com>