doc/getting-started: add null_blk tutorial
This PR adds a tutorial that will guide first-time
readers to created a zoned null block device.

This is the simplest of the three scenarios that
we will add to the guide. The other two, in order
of increasing complexity, are (1) emulated devices
(either qemu ZNS or scsi tcmu-runner) and (2) real
hardware.

This is part of an initiative to reorganize the Getting
Started page so that it turns beginners into adepts as
quickly as possible.

Signed-off-by: Zac Dover <zac.dover@gmail.com>