doc/getting-started: update zbd-emulation.md
This commit makes the text in the "Zoned Block Device Emulation
with null_blk" section a little more elegant. I have removed some
pleonasm and a couple of instances of what could be taken for
marketing-speak. This commit is the result of my combination of
my understanding of English and my improed understanding of what
this text means to communicate about emulated zoned block devices.

Signed-off-by: Zac Dover <zac.dover@gmail.com>