Extract file systems into their own main section
The file system page is very long and it can be hard for the
reader to appreciate each of the file systems that are described.

Therefore the single file system section are broken into btrfs,
f2fs, and zonefs together with an overview section that also
mentions the XFS and ext4 file systems.

Signed-off-by: Matias Bjørling <matias.bjorling@wdc.com>