Reword ZNS SSD to SSD with Zoned Namespace support
A ZNS SSD is a way to describe an SSD that supports Zoned Namespaces.
However, when using the abbrevation, it might be understood as the
SSD only supports ZNS, and is not a conventional SSD. However, NVMe
allows both conventional and zoned namespaces to co-exist on the same
device. To avoid the misunderstanding, rework the text that uses the
ZNS SSD abbrevation to be SSD with Zoned Namespace support, which
better signifies that its a feature of a specific SSD and doesn't
limit it also to be a conventional SSD.

Signed-off-by: Matias Bjørling <matias.bjorling@wdc.com>